requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
ta>=0.11.0
schedule>=1.2.0
python-dotenv>=1.0.0
